Separated 32 and 64 bit constructors and assignments for String.
git-svn-id: http://yate.null.ro/svn/yate/trunk@5611 acf43c95-373e-0410-b603-e72c3f656dc1

+++ engine/Engine.cpp
@@ -1500,7 +1500,7 @@ int Engine::run()
 	XDebug(DebugAll,"Sleeping for %ld",t);
 	Thread::usleep(t);
 	Message* m = new Message("engine.timer",0,true);
-	m->addParam("time",String((int)m->msgTime().sec()));
+	m->addParam("time",String(m->msgTime().sec()));
 	if (nodeName())
 	    m->addParam("nodename",nodeName());
 	if (s_haltcode == -1) {
