+++ libs/yscript/jsobjects.cpp
@@ -565,7 +565,7 @@ bool JsArray::runNative(ObjList& stack, const ExpOperation& oper, GenObject* con
 
 	NamedString* last = 0;
 	while (!last) {
-	    last = params().getParam(String((int)--m_length));
+	    last = params().getParam(String(--m_length));
 	    if (m_length == 0)
 		break;
 	}
@@ -598,7 +598,7 @@ bool JsArray::runNative(ObjList& stack, const ExpOperation& oper, GenObject* con
 	JsArray* array = new JsArray(mutex());
 	// copy this array
 	for (long int i = 0; i < m_length; i++)
-	    array->params().addParam(params().getParam(String((int)i)));
+	    array->params().addParam(params().getParam(String(i)));
 	array->setLength(length());
 	// add parameters (JsArray of JsObject)
 	for (long int i = oper.number(); i; i--) {
@@ -609,13 +609,13 @@ bool JsArray::runNative(ObjList& stack, const ExpOperation& oper, GenObject* con
 	    JsArray* ja = (JsArray*)obj->getObject(YATOM("JsArray"));
 	    if (ja) {
 		for (long int i = 0; i < ja->length(); i++)
-		    array->params().addParam(String((int)(i + array->length())),ja->params().getValue(String((int)i)));
+		    array->params().addParam(String(i + array->length()),ja->params().getValue(String(i)));
 		array->setLength(array->length() + ja->length());
 	    }
 	    else {
 		JsObject* jo = (JsObject*)obj->getObject(YATOM("JsObject"));
 		if (jo) {
-		    array->params().addParam(new NamedPointer(String((unsigned int)array->length()),jo));
+		    array->params().addParam(new NamedPointer(String(array->length()),jo));
 		    array->setLength(array->length() + 1);
 		    jo->ref();
 		}
@@ -640,7 +640,7 @@ bool JsArray::runNative(ObjList& stack, const ExpOperation& oper, GenObject* con
 	}
 	String result;
 	for (long int i = 0; i < length(); i++)
-	    result.append(params()[String((int)i)],separator);
+	    result.append(params()[String(i)],separator);
 	ExpEvaluator::pushOne(stack,new ExpOperation(result));
     }
     else if (oper.name() == YSTRING("reverse")) {
@@ -651,10 +651,10 @@ bool JsArray::runNative(ObjList& stack, const ExpOperation& oper, GenObject* con
 	String separator = ",";
 	String toCopy;
 	for (long int i = 0; i < length(); i++)
-	    toCopy.append(params()[String((int)i)],separator);
+	    toCopy.append(params()[String(i)],separator);
 	reversed.copyParams(params(),toCopy);
 	for (long int i = length(); i; i--)
-	    params().setParam(String((int)(length() - i)),reversed.getValue(String((int)(i - 1))));
+	    params().setParam(String(length() - i),reversed.getValue(String(i - 1)));
     }
     else if (oper.name() == YSTRING("shift")) {
 	// Removes the first element from an array and returns that element
@@ -674,8 +674,8 @@ bool JsArray::runNative(ObjList& stack, const ExpOperation& oper, GenObject* con
 	    ExpEvaluator::pushOne(stack,new ExpOperation(params().getValue("0")));
 	    // shift : value n+1 becomes value n
 	    for (long int i = 0; i < length() - 1; i++)
-		params().setParam(String((int)i),params().getValue(String((int)i + 1)));
-	    params().clearParam(String((int)(length() - 1)));
+		params().setParam(String(i),params().getValue(String(i + 1)));
+	    params().clearParam(String(length() - 1));
 	    setLength(length() - 1);
 	}
     }
@@ -693,7 +693,7 @@ bool JsArray::runNative(ObjList& stack, const ExpOperation& oper, GenObject* con
 	// shift array
 	long shift = oper.number();
 	for (long int i = length(); i; i--)
-	    params().setParam(String((int)(i - 1 + shift)),params().getValue(String((int)(i - 1))));
+	    params().setParam(String(i - 1 + shift),params().getValue(String(i - 1)));
 	for (long int i = shift; i; i--) {
 	    ExpOperation* op = popValue(stack,context);
 	    ExpWrapper* obj = YOBJECT(ExpWrapper,op);
@@ -703,8 +703,8 @@ bool JsArray::runNative(ObjList& stack, const ExpOperation& oper, GenObject* con
 	    if (!jo)
 		continue;
 	    jo->ref();
-	    params().clearParam(String((int)(i - 1)));
-	    params().setParam(new NamedPointer(String((int)(i - 1)),jo));
+	    params().clearParam(String(i - 1));
+	    params().setParam(new NamedPointer(String(i - 1),jo));
 	    TelEngine::destruct(op);
 	}
 	setLength(length() + shift);
@@ -724,7 +724,7 @@ bool JsArray::runNative(ObjList& stack, const ExpOperation& oper, GenObject* con
 	String separator = ",";
 	String result;
 	for (long int i = 0; i < length(); i++)
-	    result.append(params()[String((int)i)],separator);
+	    result.append(params()[String(i)],separator);
 	ExpEvaluator::pushOne(stack,new ExpOperation(result));
     } else if (oper.name() == YSTRING("indexOf")) {
 	ObjList args;
@@ -797,7 +797,7 @@ bool JsArray::runNativeSlice(ObjList& stack, const ExpOperation& oper, GenObject
     // TODO
     //JsArray* slice = new JsArray(mutex());
     for (long int i = begin; i < end; i++) {
-//	slice->params().addParam(new NamedString(String((int)(i - begin),
+//	slice->params().addParam(new NamedString(String(i - begin),
     }
     return true;
 }
